+++ Card/src/androidTest/java/com/braintreepayments/api/CardClientTest.java
@@ -173,7 +173,6 @@ public class CardClientTest {
                 assertEquals("Unknown", cardNonce.getCardType());
                 assertEquals("", cardNonce.getLastFour());
                 assertEquals("", cardNonce.getLastTwo());
-                assertNotNull(cardNonce.getThreeDSecureInfo());
                 assertFalse(cardNonce.isDefault());
                 assertNotNull(cardNonce.getString());
 
@@ -310,7 +309,6 @@ public class CardClientTest {
             assertEquals(BinData.UNKNOWN, cardNonce.getBinData().getIssuingBank());
             assertEquals(BinData.UNKNOWN, cardNonce.getBinData().getCountryOfIssuance());
             assertEquals(BinData.UNKNOWN, cardNonce.getBinData().getProductId());
-            assertFalse(cardNonce.getThreeDSecureInfo().wasVerified());
 
             countDownLatch.countDown();
         });